ai_engineer_assistant_project/backend/requirements.txt

fastapi
uvicorn
openai
anthropic
google-generativeai
python-dotenv
